faster bloch_redfield_tensor
